updated heartbeat layout to match everything
