+++ src/interfaces/system_interface/CMakeLists.txt
@@ -20,8 +20,8 @@ include_directories(include)
 
 add_executable(system src/system.cpp src/agent_tracker.cpp src/system_node.cpp)
 add_executable(occupancy_grid src/occupancy_grid.cpp)
-add_executable(heartbeat_server src/heartbeat_server.cpp)
-add_executable(heartbeat_client src/heartbeat_client.cpp)
+add_executable(heartbeat_server src/heartbeat_server.cpp src/heartbeat_server_node.cpp)
+add_executable(heartbeat_client src/heartbeat_client.cpp src/heartbeat_client_node.cpp)
 
 ament_target_dependencies(system rclcpp sensor_msgs px4_msgs utils domain_bridge std_msgs)
 ament_target_dependencies(occupancy_grid nav_msgs geometry_msgs utils)
