+++ src/interfaces/system_interface/src/heartbeat_client.cpp
@@ -3,55 +3,31 @@
  * About: Sender (client) node for heartbeat
  */
 
-#include <rclcpp/rclcpp.hpp>
+#include "system_interface/heartbeat_client.hpp"
 
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#include <string>
-
-using namespace std::chrono;
-using namespace std::chrono_literals;
-
-
-class HeartbeatClient : public rclcpp::Node
+HeartbeatClient::HeartbeatClient() : Node("heartbeat_client")
 {
-public:
-  HeartbeatClient() : Node("heartbeat_client")
-  {
-    const char* group_ = "226.1.1.1";
+  const char* group_ = "226.1.1.1";
 
-    socket_fd_ = socket(AF_INET, SOCK_DGRAM, 0);
+  socket_fd_ = socket(AF_INET, SOCK_DGRAM, 0);
 
-    if (socket_fd_ < 0)
-      {
-	RCLCPP_ERROR(this->get_logger(), "Error connecting to socket");
-      }
+  if (socket_fd_ < 0)
+    {
+      RCLCPP_ERROR(this->get_logger(), "Error connecting to socket");
+    }
 
-    memset((char*) &address_, 0, sizeof(address_));
+  memset((char*) &address_, 0, sizeof(address_));
 
-    address_.sin_family = AF_INET;
-    address_.sin_addr.s_addr = inet_addr(group_);
-    address_.sin_port = htons(5005);
+  address_.sin_family = AF_INET;
+  address_.sin_addr.s_addr = inet_addr(group_);
+  address_.sin_port = htons(5005);
 
-    local_interface_.s_addr = inet_addr("203.106.93.94");
-    setsockopt(socket_fd_, IPPROTO_IP, IP_MULTICAST_IF, (char*) &local_interface_, sizeof(local_interface_));
+  local_interface_.s_addr = inet_addr("203.106.93.94");
+  setsockopt(socket_fd_, IPPROTO_IP, IP_MULTICAST_IF, (char*) &local_interface_, sizeof(local_interface_));
 
-    timer_ = this -> create_wall_timer(10000ms, std::bind(&HeartbeatClient::sendHeartbeat, this));
-  }
+  timer_ = this -> create_wall_timer(10000ms, std::bind(&HeartbeatClient::sendHeartbeat, this));
+}
   
-private:
-  std::string group_;
-  int socket_fd_;
-
-  struct sockaddr_in address_;
-  struct in_addr local_interface_;
-
-  rclcpp::TimerBase::SharedPtr timer_;
-
-  void sendHeartbeat();
-};
 
 void HeartbeatClient::sendHeartbeat()
 {
@@ -63,15 +39,4 @@ void HeartbeatClient::sendHeartbeat()
     {
       RCLCPP_ERROR(this->get_logger(), "Error sending heartbeat");
     }
-}
-
-
-int main(int argc, char *argv[])
-{
-  rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<HeartbeatClient>());
-
-  rclcpp::shutdown();
-  
-  return 0;
 }
